+++ docs/src/examples/QasGrabbable/WithScroll.vue
@@ -1,7 +1,7 @@
 <template>
   <qas-grabbable use-scroll-bar>
-    <qas-box v-for="i in 20" :key="i" class="bg-grey-5 cursor-pointer q-ma-xs" @click="onClick(i)">
-      {{ i }}
+    <qas-box v-for="item in 20" :key="item" class="bg-grey-4 cursor-pointer q-ma-sm text-center" @click="onClick(item)">
+      Item {{ item }}
     </qas-box>
   </qas-grabbable>
 </template>
@@ -9,8 +9,8 @@
 <script>
 export default {
   methods: {
-    onClick (index) {
-      alert(`Clicado no box ${index}`)
+    onClick (item) {
+      alert(`Clicado no box ${item}`)
     }
   }
 }